Can now restart task sets

+++ src/servers/web/client/src/app/components/taskset-details/taskset-details.component.ts
@@ -1,9 +1,12 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ITaskSet } from '../../../../../../../database/models/taskSet';
 import { TasksetService } from 'src/app/services/taskset.service';
 import { DialogService } from 'src/app/services/dialog.service';
 import { getErrorMessage } from 'src/app/classes/utils';
+import { LacunaMaterialTableComponent } from 'lacuna-mat-table';
+import { ITask } from '../../../../../../../database/models/task';
+import { TaskListComponent } from '../task-list/task-list.component';
 
 @Component({
   selector: 'app-taskset-details',
@@ -13,6 +16,7 @@ import { getErrorMessage } from 'src/app/classes/utils';
 export class TasksetDetailsComponent implements OnInit {
 
   taskset: ITaskSet;
+  tasksetId: string;
   loading: boolean = false;
 
   constructor(
@@ -23,9 +27,15 @@ export class TasksetDetailsComponent implements OnInit {
   ) { }
 
   ngOnInit() {
-    let tasksetId = this.route.snapshot.params['tasksetId'];
+    this.tasksetId = this.route.snapshot.params['tasksetId'];
+    this.loadTaskset();
+  }
+
+  @ViewChild(TaskListComponent, { static: false }) taskList: TaskListComponent;
+
+  loadTaskset() {
     this.loading = true;
-    this.tasksetService.get(tasksetId).subscribe(taskset => {
+    this.tasksetService.get(this.tasksetId).subscribe(taskset => {
       this.taskset = taskset;
       this.loading = false;
     }, err => {
@@ -36,15 +46,54 @@ export class TasksetDetailsComponent implements OnInit {
     });
   }
 
+  cancel() {
+    this.dialogService.confirm(
+      'Are you sure you want to delete this task set? This action is irreversible!',
+      'Delete task set?'
+    ).subscribe(confirm => {
+      if (!confirm) return;
+      this.loading = true;
+      this.tasksetService.cancel(this.taskset._id).subscribe(ts => {
+        this.taskset = ts;
+        this.loading = false;
+      }, err => {
+        console.error(err);
+        this.dialogService.alert(getErrorMessage(err), 'Could not cancel the Taskset');
+      });
+    });
+  }
+
   delete() {
-    this.loading = true;
-    this.tasksetService.delete(this.taskset._id).subscribe(() => {
-      this.loading = false;
-      this.router.navigate(['..'], { relativeTo: this.route });
-    }, err => {
-      console.error(err);
-      this.dialogService.alert(getErrorMessage(err), 'Could not delete the Taskset').subscribe(() => {
+    this.dialogService.confirm(
+      'Are you sure you want to delete this task set? This action is irreversible!',
+      'Delete task set?'
+    ).subscribe(confirm => {
+      if (!confirm) return;
+      this.loading = true;
+      this.tasksetService.delete(this.taskset._id).subscribe(() => {
+        this.loading = false;
         this.router.navigate(['..'], { relativeTo: this.route });
+      }, err => {
+        console.error(err);
+        this.dialogService.alert(getErrorMessage(err), 'Could not delete the Taskset');
+      });
+    });
+  }
+
+  restart() {
+    this.dialogService.confirm(
+      'Are you sure you want to restart this task set? This action is irreversible!',
+      'Restart task set?'
+    ).subscribe(confirm => {
+      if (!confirm) return;
+      this.loading = true;
+      this.tasksetService.restart(this.taskset._id).subscribe(() => {
+        this.loading = false;
+        this.loadTaskset();
+        this.taskList.refresh();
+      }, err => {
+        console.error(err);
+        this.dialogService.alert(getErrorMessage(err), 'Could not restart the Taskset');
       });
     });
   }
@@ -60,19 +109,6 @@ export class TasksetDetailsComponent implements OnInit {
 
   graphs() {
     this.router.navigate(['graphs'], { relativeTo: this.route });
-  }
-
-  cancel() {
-    this.loading = true;
-    this.tasksetService.cancel(this.taskset._id).subscribe(ts => {
-      this.taskset = ts;
-      this.loading = false;
-    }, err => {
-      console.error(err);
-      this.dialogService.alert(getErrorMessage(err), 'Could not cancel the Taskset').subscribe(() => {
-        this.router.navigate(['..'], { relativeTo: this.route });
-      });
-    });
   }
 
   export() {
